requirements.txt
environ-config==22.1.0
Flask==2.2.2
requests==2.28.1
structlog==22.3.0
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-grpc
